test(series-monitoring): align regression with test framework
